+++ controllers/sys/dict_data.go
@@ -1,11 +1,12 @@
 package sys
 
 import (
+	"github.com/astaxie/beego/validation"
 	"github.com/syyongx/php2go"
 	"go-cms/controllers"
+	"encoding/json"
 	"go-cms/models"
-	"github.com/astaxie/beego/validation"
-	"go-cms/pkg/e"
+    "go-cms/pkg/e"
 	"log"
 )
 
@@ -17,51 +18,73 @@ func (c *DictDataController) Prepare() {
 
 }
 
+/**
+获取列表数据
+ */
 func (c *DictDataController) Index() {
-    if c.Ctx.Input.IsPost() {
-		page, _ := c.GetInt("page",1)
-		limit, _ := c.GetInt("limit",10)
-		
+	if c.Ctx.Input.IsPost() {
 		model := models.NewDictData()
+		
+		data := c.Ctx.Input.RequestBody
+		//json数据封装到user对象中
+		err := json.Unmarshal(data, &model)
+		if err != nil {
+			c.JsonResult(e.ERROR, err.Error())
+		}
 		
 		dataMap := make(map[string]interface{}, 0)
 		
 		if !php2go.Empty(model.DictId) {
 			dataMap["dict_id"] = model.DictId
 		}
-	
-	    if !php2go.Empty(model.DictLabel) {
-		    dataMap["dict_label"] = model.DictLabel
-	    }
 		
-	    if !php2go.Empty(model.Status) {
-		    dataMap["status"] = model.Status
-	    }
-		
+		if !php2go.Empty(model.DictLabel) {
+			dataMap["dict_label"] = model.DictLabel
+		}
+
+		//开始时间
 		if !php2go.Empty(model.StartTime) {
 			dataMap["start_time"] = model.StartTime
 		}
+		
+		//结束时间
 		if !php2go.Empty(model.EndTime) {
 			dataMap["end_time"] = model.EndTime
 		}
 		
+		//状态
+		if !php2go.Empty(model.Status) {
+			dataMap["status"] = model.Status
+		}
+		
 		var orderBy string = "created_at DESC"
 		
-		result, count,err := models.NewDictData().FindByMap((page-1)*limit, limit, dataMap,orderBy)
+		result, count,err := models.NewDictData().FindByMap((model.Page-1)*model.PageSize, model.PageSize, dataMap,orderBy)
 		if err != nil{
 			c.JsonResult(e.ERROR, "获取数据失败")
 		}
-		c.JsonResult(e.SUCCESS, "ok", result, count)
+		c.JsonResult(e.SUCCESS, "ok", result, count, model.Page, model.PageSize)
 	}
 }
 
+/**
+创建数据
+*/
 func (c *DictDataController) Create() {
 	if c.Ctx.Input.IsPost() {
 		model := models.NewDictData()
-		//1.压入数据
+        data := c.Ctx.Input.RequestBody
+		//1.压入数据 json数据封装到对象中
+		
+		err := json.Unmarshal(data, model)
+		if err != nil {
+			c.JsonResult(e.ERROR, err.Error())
+		}
+		
 		if err := c.ParseForm(model); err != nil {
 			c.JsonResult(e.ERROR, "赋值失败")
 		}
+
 		//2.验证
 		valid := validation.Validation{}
 		if b, _ := valid.Valid(model); !b {
@@ -70,25 +93,35 @@ func (c *DictDataController) Create() {
 			}
 			c.JsonResult(e.ERROR, "验证失败")
 		}
+
 		//3.插入数据
 		if _, err := model.Create(); err != nil {
 			c.JsonResult(e.ERROR, "创建失败")
 		}
 		c.JsonResult(e.SUCCESS, "添加成功")
 	}
-	c.Data["vo"] = models.DictData{}
-	c.TplName = c.ADMIN_TPL + "dictData/create.html"
 }
 
+/**
+更新数据
+*/
 func (c *DictDataController) Update() {
-	if c.Ctx.Input.IsPost() {
-		id, _ := c.GetInt("id")
-		model, _ := models.NewDictData().FindById(id)
-		//1
-		if err := c.ParseForm(&model); err != nil {
-			c.JsonResult(e.ERROR, "赋值失败")
+	if c.Ctx.Input.IsPut() {
+		model := models.NewDictData()
+		data := c.Ctx.Input.RequestBody
+		//json数据封装到对象中
+		
+		err := json.Unmarshal(data, model)
+		
+		if err != nil {
+			c.JsonResult(e.ERROR, err.Error())
 		}
-		//2
+		
+		post, err := models.NewDictData().FindById(model.Id)
+		if err != nil||php2go.Empty(post) {
+			c.JsonResult(e.ERROR, "没找到数据")
+		}
+		
 		valid := validation.Validation{}
 		if b, _ := valid.Valid(model); !b {
 			for _, err := range valid.Errors {
@@ -96,7 +129,7 @@ func (c *DictDataController) Update() {
 			}
 			c.JsonResult(e.ERROR, "验证失败")
 		}
-		//3
+		
 		if _, err := model.Update(); err != nil {
 			c.JsonResult(e.ERROR, "修改失败")
 		}
@@ -104,23 +137,41 @@ func (c *DictDataController) Update() {
 	}
 }
 
+/**
+删除数据
+*/
 func (c *DictDataController) Delete() {
-    model := models.NewDictData()
-	id, _ := c.GetInt("id")
-	model.Id = id
-	if err := model.Delete(); err != nil {
-		c.JsonResult(e.ERROR, "删除失败")
+    if c.Ctx.Input.IsDelete() {
+		model := models.NewDictData()
+		data := c.Ctx.Input.RequestBody
+		//json数据封装到user对象中
+		
+		err := json.Unmarshal(data, model)
+		
+		if err != nil {
+			c.JsonResult(e.ERROR, err.Error())
+		}
+		
+		post, err := models.NewDictData().FindById(model.Id)
+		if err != nil||php2go.Empty(post) {
+			c.JsonResult(e.ERROR, "没找到数据")
+		}
+		
+		if err := model.Delete(); err != nil {
+			c.JsonResult(e.ERROR, "删除失败")
+		}
+		c.JsonResult(e.SUCCESS, "删除成功")
 	}
-	c.JsonResult(e.SUCCESS, "删除成功")
 }
 
 func (c *DictDataController) BatchDelete() {
+	model := models.NewDictData()
+
 	var ids []int
 	if err := c.Ctx.Input.Bind(&ids, "ids"); err != nil {
 		c.JsonResult(e.ERROR, "赋值失败")
 	}
-	
-	model := models.NewDictData()
+
 	if err := model.DelBatch(ids); err != nil {
 		c.JsonResult(e.ERROR, "删除失败")
 	}
